refactor: F.S.D
