+++ apps/community/src/entities/guide/ui/GuideModal.tsx
